==> 1. Allocate data.ipynb <==
## Allocate data - 크롤링으로

from selenium.webdriver import Chrome
import requests
from sklearn.feature_extraction.text import CountVectorizer #feature : 변수  / CountVectorizer : TDM을 만들어준다.
import time
import numpy
import operator
from konlpy.tag import Komoran
from urllib.parse import quote_plus
import lxml.html
import pickle
import re

import pickle

## 자소서 url을 가져오자
apply_url = []

for page in range(1, 2): # page 1 ~ 229
    url = 'http://www.jobkorea.co.kr/starter/PassAssay?FavorCo_Stat=0&Pass_An_Stat=0&OrderBy=0&EduType=0&WorkType=0&isSaved=0&Page={}'.format(page) # 페이지별로 url 바꿔주기
    res = requests.get(url) # url을 불러온다
    root = lxml.html.fromstring(res.text) # 분석
    for url_code in root.cssselect('ul p a'):
        apply_url.append('http://www.jobkorea.co.kr' + url_code.attrib['href'])
        
## 15min 소요

apply_url[1:10]

## pickle에 저장하기
# with open('apply_url.txt', 'wb') as f:
#     pickle.dump(apply_url, f)
# f.close

## pickle load 하기
with open('apply_url.txt', 'rb') as f:
    apply_url = pickle.load(f)

apply_url[1:10]

len(apply_url) # 4850개의 자소서

win_apply = [] # 4850개의 자소서를 가져오자

for url in apply_url: # 4850개의 합격자소서 url 가져오기
    res = requests.get(url)
    root = lxml.html.fromstring(res.text)
    for content in root.cssselect('dd div.tx'):
        win_apply.append(content.text_content().replace("아쉬운점 1", "").replace("좋은점 1", "").replace("아쉬운점 2", "").replace("좋은점 2", "")
                        .replace("아쉬운점 3", "").replace("좋은점 3", "").replace("아쉬운점 4", "").replace("좋은점 4", "").replace("아쉬운점 5", "")
                         .replace("좋은점 5", "").strip())

# 역 1h 소요

win_apply[0:5]

# 글자수, Byte 제거
import re

p = re.compile('\글자수.*?\Byte')

for i in range(len(win_apply)):
    real_text = re.sub(p, "", win_apply[i])
    win_apply[i] = real_text

# "" 안 소제목 없애기
import re

p = re.compile('\".*?\"')

for i in range(len(win_apply)):
    real_text = re.sub(p, "", win_apply[i])
    win_apply[i] = real_text

# 공백, \r, 숫자) 없애기
win_apply2 = []

for i in win_apply:
    win_apply2.append(i.replace("\r","").replace("1)","").replace("2)","").replace("3)","")
                      .replace("4)","").replace("5)","").replace("6)","").strip())

win_apply = win_apply2

win_apply[0:4]

## 문장 수 구하기
count_sentence = 0

for i in win_apply:
    a = float(i.count("다."))
    count_sentence = count_sentence + a

win_apply[0]

count_sentence # 185224 문장

len(win_apply) #18605개의 자소서 항목

import pickle

## pickle에 저장하기
# with open('win_apply.txt', 'wb') as f:
#     pickle.dump(win_apply, f)
# f.close

## pickle load 하기
with open('win_apply.txt', 'rb') as f:
    win_apply = pickle.load(f)

win_apply[0:3]

### 형태소 분석(Term-Document Matrix)

## 형태소 분석기
tagger = Komoran()

## 명사추출기 함수
def get_noun(text):
    nouns = tagger.nouns(text)
    return [n for n in nouns if len(n) >= 1]  # 명사만 추출

## tokenizer
cv = CountVectorizer(tokenizer=get_noun)

## tdm 만들기
# 5min 소요
tdm = cv.fit_transform(win_apply)

## 단어 빈도 구하기
words = cv.get_feature_names()
count_mat = tdm.sum(axis=0)

count = numpy.squeeze(numpy.asarray(count_mat))

word_count = list(zip(words, count))

word_count = sorted(word_count, key=operator.itemgetter(1), reverse=True) # 역순으로 정렬

#word_count
