==> 2. Data_preprocess 1.ipynb <==
### 시작! : 전처리를 해보자

import requests
from sklearn.feature_extraction.text import CountVectorizer
from gensim.models.word2vec import Word2Vec
from keras.preprocessing.text import Tokenizer
import time
import numpy as np
import operator
from konlpy.tag import Komoran
from urllib.parse import quote_plus
import lxml.html
import pickle
import re
import keras
import pandas as pd

## pickle load 하기
with open('win_apply.txt', 'rb') as f:
    win_apply = pickle.load(f)

win_apply[0:2]

# sentence 만들기
sentence = []
for i in range(0, len(win_apply)):
    a = win_apply[i].split(".")
    for j in range(0, len(a)):
        b = a[j]
        sentence.append(b)

## 명사추출
tagger = Komoran()

def get_noun(text):
    nouns = tagger.nouns(text)
    return [n for n in nouns if len(n) >= 2]  # 명사만 추출

words = []
for i in range(0, len(sentence)):
    a = get_noun(sentence[i])
    words.append(a)

print("문장수 :",len(sentence))
print("명사 추출 단어 수 :",len(words))

# 전처리
none_index = []
for i in range(0, len(words)):
    if len(words[i]) == 0:
        none_index.append(i)

none_index.reverse()
        
for j in none_index:
    del sentence[j]
    del words[j]

print("값이 없는 리시트 삭제 후")
print("문장수 :",len(sentence))
print("명사 추출 단어 수 :",len(words))

max_index = []
for k in range(0, len(words)):
    if len(words[k]) > 42:
        max_index.append(k)

max_index.reverse()

for l in max_index:
    del words[l]
    del sentence[l]

print("\n")
print("명사 과다 추출된 리스트 삭제 후")
print("문장수 :",len(sentence))
print("명사 추출 단어 수 :",len(words))

### WORD - 임베딩 벡터를 만들어보자

# embedding
embedding_model = Word2Vec(words,
                           size=200,  # 100차원의 벡터로 바꾼다
                           window = 2,  # 주변 단어 앞 뒤로 두개까지
                           min_count = 1,  # 출현 빈도의 최소값, 나머진 삭제
                           workers = 4,  # CPU는 쿼드코어를 사용
                           iter = 100, # 100번 학습
                           sg = 1 # 분석방법론은 CBOW와 Skip-Gram 중 후자 선택
                          )
print(embedding_model)

# 결과 확인
print(embedding_model.most_similar(positive=["장학금"],
                                   topn=10) # 단어와 가장 비슷한(유사도가 큰) 100개 단어 출력
     )

### Image 벡터 처럼 키워드를 임베딩 벡터로 바꾸어 보장

# words에서 키워드 하나만 살리기
for i in range(0, len(words)):
    words[i] = words[i][0]

words[0:5]

# pickle에 저장하기
# with open('words_1_dic.txt', 'wb') as f:
#     pickle.dump(words, f)
# f.close

# 데이터 로드하기
with open('words.txt', 'rb') as f:
    words= pickle.load(f)
f.close

# 워드 벡터 만들기
for i in range(0, len(words)):
    words[i] = embedding_model[words[i]]

# pickle에 저장하기
# with open('words_vec_200.txt', 'wb') as f:
#     pickle.dump(words, f)
# f.close

with open('words_vec_200.txt', 'rb') as f:
    words = pickle.load(f)
f.close

with open('sentence.txt', 'wb') as f:
    pickle.dump(sentence, f)
f.close

# 완성~
print(len(words))
print(len(sentence))

### 문단 전처리를 합시다!

## pickle load 하기
with open('win_apply.txt', 'rb') as f:
    para = pickle.load(f)
    
# 문단에 대해 명사를 추출
tagger = Komoran()

def get_noun(text):
    nouns = tagger.nouns(text)
    return [n for n in nouns if len(n) >= 2]  # 명사만 추출

words = []
for i in range(0, len(para)):
    a = get_noun(para[i])
    words.append(a)

 # words 추출된 단어 개수 확인
words_count = []
for i in range(0,len(words)):
    words_count.append(len(words[i]))

# 인덱스 화
words_keyword_dic = {w: i for i, w in enumerate(words_count)}

print(len(words))
print(len(para))

"""
1. 300자 이하인 자소서를 제거
2. 모든 명사를 추출 및 확인
3. word2vec을 통해서 벡터화
4. padding을 통해서 맞춰주기 → 함수 만들기
5. 모델 돌리기
"""

"""
## 임베딩 벡터 만들기 ## 
데이터 전처리
words : 명사 추출 데이터
para : 문항 데이터
para_length : 문항의 최소 글자 수
"""

para_length = 400

index_over = []
for i in range(0, len(para)):
    if len(para[i]) < para_length:
        index_over.append(i)
    
index_over.reverse() # 역순으로
    
# index_over 값을 데이터에서 제거
for i in index_over:
    del words[i], para[i]
    
# padding을 위해서(나머지 자리 0으로 채워주기 위해서) 단어의 최고 자리 구하기
# words 추출된 단어 개수 확인
words_count = []
for i in range(0,len(words)):
    words_count.append(len(words[i]))
    
words_length_max = max(words_count) # max_length 설정
    
# Words Word2Vec
embedding_model = Word2Vec(words,
                           size = 100,  # 200차원의 벡터로 바꾼다
                           window = 2,  # 주변 단어 앞 뒤로 두개까지
                           min_count = 1,  # 출현 빈도의 최소값, 나머진 삭제
                           workers = 4,  # CPU는 쿼드코어를 사용
                           iter = 100, # 100번 학습
                           sg = 1 # 분석방법론은 CBOW와 Skip-Gram 중 후자 선택
                          )    


# 단어를 벡터 값으로 바꾸기
words_vec = []
for i in range(0, len(words)):
    words_vec_before = []
    
    for j in range(0, len(words[i])):
        a = embedding_model[words[i][j]]
        for k in range(0, len(a)):
            words_vec_before.append(a[k])
                
    words_vec.append(words_vec_before)

print(len(words_vec))
print(len(words_vec[0]))

"""
dense vector shape fittning
벡터의 shape 맞추기! = padding : 0으로 채우자
"""
words_vec_length_count = []
for i in range(0, len(words_vec)):
    words_vec_length_count.append(len(words_vec[i]))

vec_max_length = max(words_vec_length_count) # 

for i in range(0, len(words_vec)):
    if len(words_vec[i]) != vec_max_length:
        for j in range(0, (vec_max_length-len(words_vec[i]))):
            words_vec[i].append(0)

len(words_vec[0])

words_vec = np.asarray(words_vec)

# pickle에 저장하기
# test0515 : 5차원 테스트용
# test0516 : 100차원 발표 결과용
with open('test0516.txt', 'wb') as f:
    pickle.dump(words_vec, f)
f.close

# pickle에 저장하기
# para0515 : 5차원 테스트
# para0516 : 100차원 발표 결과용
with open('para0516.txt', 'wb') as f:
    pickle.dump(para, f)
f.close

## 0523

## pickle load 하기
with open('Keyword0523.txt', 'rb') as f:
    words = pickle.load(f)
f.close()
    
## pickle load 하기
with open('para0523.txt', 'rb') as f:
    para = pickle.load(f)
f.close()

print(words[0:1])
print(para[0:1])

"""
1. 추출 단어가 30개 미만, 200개 초과인거 삭제
2. 모든 명사를 추출 및 확인
3. word2vec을 통해서 벡터화 200차원
4. padding을 통해서 맞춰주기 → 함수 만들기
5. 모델 돌리기
"""

"""
## 임베딩 벡터 만들기 ## 
데이터 전처리
words : 명사 추출 데이터
para : 문항 데이터
para_length : 문항의 최소 글자 수
"""
"""
[0523 keyword, paragraph 파일]
추출 단어가 30개 미만, 200개 초과인거 삭제
"""

"""
이 부분은 이미 진행되었으므로 skip
para_length = 400

index_over = []
for i in range(0, len(para)):
    if len(para[i]) < para_length:
        index_over.append(i)
    
index_over.reverse() # 역순으로
    
# index_over 값을 데이터에서 제거
for i in index_over:
    del words[i], para[i]

"""    
# padding을 위해서(나머지 자리 0으로 채워주기 위해서) 단어의 최고 자리 구하기
# words 추출된 단어 개수 확인
words_count = []
for i in range(0,len(words)):
    words_count.append(len(words[i]))
    
words_length_max = max(words_count) # max_length 설정
    
# Words Word2Vec
embedding_model = Word2Vec(words,
                           size = 200,  # 200차원의 벡터로 바꾼다
                           window = 2,  # 주변 단어 앞 뒤로 두개까지
                           min_count = 1,  # 출현 빈도의 최소값, 나머진 삭제
                           workers = 4,  # CPU는 쿼드코어를 사용
                           iter = 100, # 100번 학습
                           sg = 1 # 분석방법론은 CBOW와 Skip-Gram 중 후자 선택
                          )    


# 단어를 벡터 값으로 바꾸기
words_vec = []
for i in range(0, len(words)):
    words_vec_before = []
    
    for j in range(0, len(words[i])):
        a = embedding_model[words[i][j]]
        for k in range(0, len(a)):
            words_vec_before.append(a[k])
                
    words_vec.append(words_vec_before)

print(len(words_vec))
print(len(words_vec[0]))

"""
dense vector shape fittning
벡터의 shape 맞추기! = padding : 0으로 채우자
"""
words_vec_length_count = []
for i in range(0, len(words_vec)):
    words_vec_length_count.append(len(words_vec[i]))

vec_max_length = max(words_vec_length_count) # 

for i in range(0, len(words_vec)):
    if len(words_vec[i]) != vec_max_length:
        for j in range(0, (vec_max_length-len(words_vec[i]))):
            words_vec[i].append(0)

len(words_vec[0])

words_vec = np.asarray(words_vec)

# pickle에 저장하기
# test0515 : 5차원 테스트용
# test0516 : 100차원 발표 결과용
# test0523 : 200차원 면담 결과용
with open('test0523.txt', 'wb') as f:
    pickle.dump(words_vec, f)

### 영어 자소서 확인

import pickle
import numpy as np

with open('para0523.txt', 'rb') as f:
    para = np.array(pickle.load(f))
        
with open('test0523.txt', 'rb') as f:
    test = np.array(pickle.load(f))

index = []
for i in range(0, len(para)):
    if (("a" in para[i]) & ("b" in para[i]) & ("c" in para[i]) & ("d" in para[i]) & ("e" in para[i]) & ("f" in para[i]) & ("g" in para[i]) & ("h" in para[i]) & ("i" in para[i]) ):
        index.append(i)

print("인덱스 개수 : ",len(index))
print("변경 전 : ", len(para))
print("변경 전 : ",len(test))

# list로 바꿔서 해당 인덱스 삭제
para = para.tolist()
test = test.tolist()
index.reverse()
for i in index:
    del para[i]
    del test[i]
    
print("변경 후 : ", len(para))
print("변경 후 : ",len(test))

with open('test0528.txt', 'wb') as f:
    pickle.dump(test, f)
    
with open('para0528.txt', 'wb') as f:
    pickle.dump(para, f)
