==> 3. TF-IDF keyword.ipynb <==
import requests
from sklearn.feature_extraction.text import CountVectorizer
from gensim.models.word2vec import Word2Vec
from keras.preprocessing.text import Tokenizer
import time
import numpy as np
import operator
from konlpy.tag import Komoran, Twitter ,Kkma
from urllib.parse import quote_plus
import lxml.html
import pickle
import re
import keras
import pandas as pd
from collections import Counter

## pickle load 하기
with open('win_apply.txt', 'rb') as f:
    win_apply = pickle.load(f)

## 명사추출
tagger = Komoran()
def get_noun(text):
    nouns = tagger.nouns(text)
    return [n for n in nouns if len(n) >= 2]  # 명사만 추출

words = []
for i in range(0, len(paragraph)):
    a = get_noun(paragraph[i])
    words.append(a)

## 기본 전처리 후 TF-iDF 사용(시간이 오래 걸리기 때문에)

with open('words0517.txt', 'rb') as f:
        #words = np.array(pickle.load(f))
        words = pickle.load(f)
f.close()

with open('paragraph0517.txt', 'rb') as f:
        paragraph = np.array(pickle.load(f))
f.close()

def get_tags(words, ntags=100):
    # konlpy의 Twitter객체
    # nouns 함수를 통해서 text에서 명사만 분리/추출
    count = Counter(words)
    # Counter객체를 생성하고 참조변수 nouns할당
    return_list = []  # 명사 빈도수 저장할 변수
    for n, c in count.most_common(ntags):
        temp = {'noun': n, 'count': c}
        return_list.append(temp)
    return return_list

import math
from textblob import TextBlob as tb

def tf(word, blob):
    return blob.words.count(word) / len(blob.words)

def n_containing(word, bloblist):
    return sum(1 for blob in bloblist if word in blob.words)

def idf(word, bloblist):
    return math.log(len(bloblist) / (1 + n_containing(word, bloblist)))

def tfidf(word, blob, bloblist):
    return tf(word, blob) * idf(word, bloblist)

#문항별 명사 list -> string으로
words_new=[]
for i in range(0,len(words)):  
    #words_string= ','.join(map(str,words[i]))
    words_string =','.join(words[i])
    words_new.append(words_string)

#문항별 명사 전부 tb화 
words_tb=[]
for i in range(0,len(words_new)):
    words_re =tb(words_new[i])
    words_tb.append(words_re)

bloblist =list(words_tb)

#수정해야할수도있음.
keyword =[]
Keyword =[]
for i, blob in enumerate(bloblist):
    print("Top words in document {}".format(i))
    scores = {word: tfidf(word, blob, bloblist) for word in blob.words}
    sorted_words = sorted(scores.items(), key=lambda x: x[1], reverse=True)
    
    Keyword.append(keyword)
    keyword=[]
    
    for word, score in sorted_words[:20]:
        keyword.append(word)
        #Keyword.append(keyword)
        print("\tWord: {}, TF-IDF: {}".format(word, round(score, 5)))
#오후 3시 28분 시작  //10분에 (200)문항  ->대략 14시간 

with open('last_keyword0522.txt', 'wb') as f:
    pickle.dump(Keyword, f)
f.close
