==> 4. Data_preprocess 2.ipynb <==
### tf-idf, 추출 순서를 맞추기

import requests
from sklearn.feature_extraction.text import CountVectorizer
from gensim.models.word2vec import Word2Vec
from keras.preprocessing.text import Tokenizer
import time
import numpy as np
import operator
from konlpy.tag import Komoran
from urllib.parse import quote_plus
import lxml.html
import pickle
import re
import keras
import pandas as pd

"""
기존의 키워드가 너무 많다고 판단하였다. 따라서 기존의 tf-idf를 통해서 나온 키워드를 사용하지만 명사의 출현 순서를 고려하여 키워드 벡터를 다시 만든다.
일단 기존의 키워드 개수 20개를 그대로 가지고 간다는 것으로 생각하자
"""

## pickle load 및 확인
with open('Keyword0523.txt', 'rb') as f:
    keyword = pickle.load(f)
    
with open('para0523.txt', 'rb') as f:
    para = pickle.load(f)
    para = list(para)

print(len(keyword))
print(len(para))
print(keyword[0])
print(para[0])

tagger = Komoran()

def get_noun(text):
    nouns = tagger.nouns(text)
    return [n for n in nouns if len(n) >= 2]  # 명사만 추출

words = []
for i in range(0, len(para)):
    a = get_noun(para[i])
    words.append(a)

new_words = []
for i in range(0, len(words)):
    a = []
    for j in range(0, len(words[i])):
        if words[i][j] in keyword[i]:
            if words[i][j] not in a:
                a.append(words[i][j])
    new_words.append(a)
print('개수는 :', len(new_words))

print(new_words[0:5])

# 실직적으로 20개 되지 않는 index들
index = []
for i in range(0, len(new_words)):
    if len(new_words[i]) != 20:
        index.append(i)
print("20개가 되지 않는 개수 :", len(index))

"""
20개씩 키워드를 정렬하였지만 20개가 아닌 것들이 무려 4730개가 되었다.
따라서 어떤 인덱스에서 최소 값과 최대 값을 알아보자
"""

# 각 index의 길이 알아보기
index_length = []
for i in range(0, len(index)):
    index_length.append(len(new_words[i]))

print("최소값 :", min(index_length))
print("최대값 :", max(index_length))

"""
길이의 최소값은 9이고 최대값은 20이다.
"""

# 20개 나오지 않은 데이터 삭제
index.reverse()
for i in index:
    del new_words[i]
    del para[i]
print(len(new_words))
print(len(para))

# Words Word2Vec
embedding_model = Word2Vec(new_words,
                           size = 100,  # 100차원의 벡터로 바꾼다
                           window = 2,  # 주변 단어 앞 뒤로 두개까지
                           min_count = 1,  # 출현 빈도의 최소값, 나머진 삭제
                           workers = 4,  # CPU는 쿼드코어를 사용
                           iter = 100, # 100번 학습
                           sg = 1 # 분석방법론은 CBOW와 Skip-Gram 중 후자 선택
                          )

print("embedding model 학습 완료!")

# 단어를 벡터 값으로 바꾸기
words_vec = []
for i in range(0, len(new_words)):
    words_vec_before = []
    
    for j in range(0, len(new_words[i])):
        a = embedding_model[new_words[i][j]]
        for k in range(0, len(a)):
            words_vec_before.append(a[k])
                
    words_vec.append(words_vec_before)
    
print("단어를 벡터로 하는 list 완료!")
    
# padding
words_vec_length_count = []
for i in range(0, len(words_vec)):
    words_vec_length_count.append(len(words_vec[i]))

vec_max_length = max(words_vec_length_count)

for i in range(0, len(words_vec)):
    if len(words_vec[i]) != vec_max_length:
        for j in range(0, (vec_max_length-len(words_vec[i]))):
            words_vec[i].append(0)
            
print("padding 완료!")
print("만들어진 dense vector의 shape :", len(words_vec), "X", len(words_vec[0]))

# pickle에 저장하기
with open('test0607.txt', 'wb') as f:
    pickle.dump(words_vec, f)
f.close

with open('para0607.txt', 'wb') as f:
    pickle.dump(para, f)
f.close

with open('keywords0607.txt', 'wb') as f:
    pickle.dump(new_words, f)
f.close
